% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCompactCSNEM.R
\name{getCompactCSNEM}
\alias{getCompactCSNEM}
\title{Infer effect matrix from NEM mixture}
\usage{
getCompactCSNEM(big.nem, n.actions, k)
}
\arguments{
\item{big.nem}{the big (nk x nk) nem}

\item{n.actions}{the number of (actual) actions n}

\item{k}{the number of contexts}
}
\value{
list of: adjacency and accessibility matrices for new nem without redundant action nodes,
effect mapping matrix, name mapping list from the big nem to the compact
}
\description{
Infer effect matrix from NEM mixture
}
