% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_constrained_nem.R
\name{infer_constrained_nem}
\alias{infer_constrained_nem}
\title{Run constrained nem inference}
\usage{
infer_constrained_nem(log_likelihood_matrix, k)
}
\arguments{
\item{log_likelihood_matrix}{A matrix where each row corresponds to an effect and each column corresponds to
an action, where each cell corresponds to the log likelihood ratio of the effect being differentially expressed
as a result of that action.}

\item{k}{The number of contexts}
}
\value{
nem object
}
\description{
Run constrained nem inference
}
