% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBigRMatrix.R
\name{makeBigRMatrix}
\alias{makeBigRMatrix}
\title{Make the big R matrix}
\usage{
makeBigRMatrix(small.R.matrix, k)
}
\arguments{
\item{small.R.matrix}{the "small" R matrix, the one where each action appears once}

\item{k}{the number of times to append the small R matrix to itself}
}
\value{
the small R matrix appended to itself k times with column names remapped to <name>_i for i in 1...k
}
\description{
Make the big R matrix
}
