% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceAdj.R
\name{reduceAdj}
\alias{reduceAdj}
\title{Reduce a repetetive big.adj to a non-redundant csnem with ancestry naming convention}
\usage{
reduceAdj(big.adj)
}
\arguments{
\item{big.adj}{a named adjacency matrix with possibly redundant nodes}
}
\value{
a named list with the new adjacency matrix and the name mapping
}
\description{
Reduce a repetetive big.adj to a non-redundant csnem with ancestry naming convention
}
