% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripSuffix.R
\name{stripSuffix}
\alias{stripSuffix}
\title{Strip everything including and following a delimiter from strings}
\usage{
stripSuffix(in.str, delim = "_")
}
\arguments{
\item{in.str}{an input string or an array of input strings}

\item{delim}{the delimiter}
}
\value{
every element of in.str taken until the first delimiter occurence exclusive
}
\description{
Strip everything including and following a delimiter from strings
}
